Basic implementation of ZSC Added temp options to class player Moved print_logo function to main Added declaration for multiplayer and bot play
